+++ CHANGELOG.md
@@ -13,5 +13,6 @@ This changelog only contains the changes that are unreleased. For changes for in
 - Issue with skins not updating after changing them [#671]
 - Switch to using id rather than updated time when checking modpacks.ch updates
 - Anaytics having wrong category on Add Mods dialog for pagination and search
+- Keep section dropdown the same when switching platform in Add Mods dialog
 
 ### Misc